do some cleanup before merging into dev
