+++ extensions/kd-animation/requirements.txt
@@ -1,23 +1,12 @@
-#diffusers==0.20.2
-#torch 
-#torchvision
-#torchaudio
-#accelerate
 clean-fid
-#colab-convert
 einops
 ftfy
 ipython
 ipywidgets
 jsonmerge
-#jupyterlab
-#jupyter_http_over_ws
 kornia
 matplotlib
-#notebook
 numexpr
-#omegaconf
-#opencv-python
 pandas
 pytorch_lightning
 resize-right
@@ -25,9 +14,6 @@ scikit-image
 scikit-learn
 timm
 torchdiffeq
-#transformers
-#safetensors
-#albumentations
 more_itertools
 devtools
 validators
@@ -36,5 +22,4 @@ open-clip-torch
 torchsde
 ninja
 pydantic
-#xformers
 imageio[ffmpeg]
